requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
Werkzeug
python-dotenv

